Split out guest metrics conversion
To keep source files focused.

Updated the signature of `PopulateCuttlefishGuest` to take a
`MetricsEventV2` reference instead of a `CuttlefishGuest` reference to
have a clean boundary where the split out conversion files handle the
submessage entirely.

Bug: 549999548
